[docs] Move TablePaginationUnstyled docs to Base
