+++ docs/data/base/components/table-pagination/TableCustomized.js
@@ -1,8 +1,10 @@
 import * as React from 'react';
 import { styled } from '@mui/system';
-import TablePaginationUnstyled from '@mui/base/TablePaginationUnstyled';
+import TablePaginationUnstyled, {
+  tablePaginationUnstyledClasses as classes,
+} from '@mui/base/TablePaginationUnstyled';
 
-function createData(name: string, calories: number, fat: number) {
+function createData(name, calories, fat) {
   return { name, calories, fat };
 }
 
@@ -22,27 +24,53 @@ const rows = [
   createData('Oreo', 437, 18.0),
 ].sort((a, b) => (a.calories < b.calories ? -1 : 1));
 
-const Root = styled('div')`
+const blue = {
+  200: '#A5D8FF',
+  400: '#3399FF',
+};
+
+const grey = {
+  50: '#F3F6F9',
+  100: '#E7EBF0',
+  200: '#E0E3E7',
+  300: '#CDD2D7',
+  400: '#B2BAC2',
+  500: '#A0AAB4',
+  600: '#6F7E8C',
+  700: '#3E5060',
+  800: '#2D3843',
+  900: '#1A2027',
+};
+
+const Root = styled('div')(
+  ({ theme }) => `
   table {
-    font-family: arial, sans-serif;
+    font-family: IBM Plex Sans, sans-serif;
+    font-size: 0.875rem;
     border-collapse: collapse;
     width: 100%;
   }
 
   td,
   th {
-    border: 1px solid #ddd;
+    border: 1px solid ${theme.palette.mode === 'dark' ? grey[800] : grey[200]};
     text-align: left;
-    padding: 8px;
+    padding: 6px;
   }
 
   th {
-    background-color: #ddd;
+    background-color: ${theme.palette.mode === 'dark' ? grey[900] : grey[100]};
   }
-`;
+  `,
+);
 
-const CustomTablePagination = styled(TablePaginationUnstyled)`
-  & .MuiTablePaginationUnstyled-toolbar {
+const CustomTablePagination = styled(TablePaginationUnstyled)(
+  ({ theme }) => `
+  & .${classes.spacer} {
+    display: none;
+  }
+
+  & .${classes.toolbar}  {
     display: flex;
     flex-direction: column;
     align-items: flex-start;
@@ -54,11 +82,26 @@ const CustomTablePagination = styled(TablePaginationUnstyled)`
     }
   }
 
-  & .MuiTablePaginationUnstyled-selectLabel {
+  & .${classes.selectLabel} {
     margin: 0;
   }
 
-  & .MuiTablePaginationUnstyled-displayedRows {
+  & .${classes.select}{
+    padding: 2px;
+    border: 1px solid ${theme.palette.mode === 'dark' ? grey[800] : grey[200]};
+    border-radius: 50px;
+    background-color: transparent;
+
+    &:hover {
+      background-color: ${theme.palette.mode === 'dark' ? grey[800] : grey[50]};
+    }
+
+    &:focus {
+      outline: 1px solid ${theme.palette.mode === 'dark' ? blue[400] : blue[200]};
+    }
+  }
+
+  & .${classes.displayedRows} {
     margin: 0;
 
     @media (min-width: 768px) {
@@ -66,15 +109,29 @@ const CustomTablePagination = styled(TablePaginationUnstyled)`
     }
   }
 
-  & .MuiTablePaginationUnstyled-spacer {
-    display: none;
+  & .${classes.actions} {
+    padding: 2px;
+    border: 1px solid ${theme.palette.mode === 'dark' ? grey[800] : grey[200]};
+    border-radius: 50px;
+    text-align: center;
   }
 
-  & .MuiTablePaginationUnstyled-actions {
-    display: flex;
-    gap: 0.25rem;
+  & .${classes.actions} > button {
+    margin: 0 8px;
+    border: transparent;
+    border-radius: 2px;
+    background-color: transparent;
+
+    &:hover {
+      background-color: ${theme.palette.mode === 'dark' ? grey[800] : grey[50]};
+    }
+
+    &:focus {
+      outline: 1px solid ${theme.palette.mode === 'dark' ? blue[400] : blue[200]};
+    }
   }
-`;
+  `,
+);
 
 export default function UnstyledTable() {
   const [page, setPage] = React.useState(0);
@@ -84,22 +141,17 @@ export default function UnstyledTable() {
   const emptyRows =
     page > 0 ? Math.max(0, (1 + page) * rowsPerPage - rows.length) : 0;
 
-  const handleChangePage = (
-    event: React.MouseEvent<HTMLButtonElement> | null,
-    newPage: number,
-  ) => {
+  const handleChangePage = (event, newPage) => {
     setPage(newPage);
   };
 
-  const handleChangeRowsPerPage = (
-    event: React.ChangeEvent<HTMLInputElement | HTMLTextAreaElement>,
-  ) => {
+  const handleChangeRowsPerPage = (event) => {
     setRowsPerPage(parseInt(event.target.value, 10));
     setPage(0);
   };
 
   return (
-    <Root sx={{ maxWidth: '100%', width: 500 }}>
+    <Root sx={{ width: 500, maxWidth: '100%' }}>
       <table aria-label="custom pagination table">
         <thead>
           <tr>
@@ -115,16 +167,17 @@ export default function UnstyledTable() {
           ).map((row) => (
             <tr key={row.name}>
               <td>{row.name}</td>
-              <td style={{ width: 160 }} align="right">
+              <td style={{ width: 120 }} align="right">
                 {row.calories}
               </td>
-              <td style={{ width: 160 }} align="right">
+              <td style={{ width: 120 }} align="right">
                 {row.fat}
               </td>
             </tr>
           ))}
+
           {emptyRows > 0 && (
-            <tr style={{ height: 41 * emptyRows }}>
+            <tr style={{ height: 34 * emptyRows }}>
               <td colSpan={3} />
             </tr>
           )}
@@ -144,7 +197,7 @@ export default function UnstyledTable() {
                 actions: {
                   showFirstButton: true,
                   showLastButton: true,
-                } as any,
+                },
               }}
               onPageChange={handleChangePage}
               onRowsPerPageChange={handleChangeRowsPerPage}
