+++ docs/data/base/components/table-pagination/TableUnstyled.tsx
@@ -1,6 +1,8 @@
 import * as React from 'react';
 import { styled } from '@mui/system';
-import TablePaginationUnstyled from '@mui/base/TablePaginationUnstyled';
+import TablePaginationUnstyled, {
+  tablePaginationUnstyledClasses as classes,
+} from '@mui/base/TablePaginationUnstyled';
 
 function createData(name: string, calories: number, fat: number) {
   return { name, calories, fat };
@@ -22,52 +24,27 @@ const rows = [
   createData('Oreo', 437, 18.0),
 ].sort((a, b) => (a.calories < b.calories ? -1 : 1));
 
-const blue = {
-  200: '#A5D8FF',
-  400: '#3399FF',
-};
-
-const grey = {
-  50: '#F3F6F9',
-  100: '#E7EBF0',
-  200: '#E0E3E7',
-  300: '#CDD2D7',
-  400: '#B2BAC2',
-  500: '#A0AAB4',
-  600: '#6F7E8C',
-  700: '#3E5060',
-  800: '#2D3843',
-  900: '#1A2027',
-};
-
-const Root = styled('div')(
-  ({ theme }) => `
+const Root = styled('div')`
   table {
-    font-family: IBM Plex Sans, sans-serif;
-    font-size: 0.875rem;
+    font-family: arial, sans-serif;
     border-collapse: collapse;
     width: 100%;
   }
 
   td,
   th {
-    border: 1px solid ${theme.palette.mode === 'dark' ? grey[800] : grey[200]};
+    border: 1px solid #ddd;
     text-align: left;
-    padding: 6px;
+    padding: 8px;
   }
 
   th {
-    background-color: ${theme.palette.mode === 'dark' ? grey[900] : grey[100]};
+    background-color: #ddd;
   }
-  `,
-);
+`;
 
-const CustomTablePagination = styled(TablePaginationUnstyled)(
-  ({ theme }) => `
-  & .MuiTablePaginationUnstyled-spacer {
-    display: none;
-  }
-  & .MuiTablePaginationUnstyled-toolbar {
+const CustomTablePagination = styled(TablePaginationUnstyled)`
+  & .${classes.toolbar} {
     display: flex;
     flex-direction: column;
     align-items: flex-start;
@@ -78,48 +55,28 @@ const CustomTablePagination = styled(TablePaginationUnstyled)(
       align-items: center;
     }
   }
-  & .MuiTablePaginationUnstyled-selectLabel {
+
+  & .${classes.selectLabel} {
     margin: 0;
   }
-  & .MuiTablePaginationUnstyled-select {
-    padding: 2px;
-    border: 1px solid ${theme.palette.mode === 'dark' ? grey[800] : grey[200]};
-    border-radius: 50px;
-    background-color: transparent;
-    &:hover {
-      background-color: ${theme.palette.mode === 'dark' ? grey[800] : grey[50]};
-    }
-    &:focus {
-      outline: 1px solid ${theme.palette.mode === 'dark' ? blue[400] : blue[200]};
-    }
-  }
-  & .MuiTablePaginationUnstyled-displayedRows {
+
+  & .${classes.displayedRows} {
     margin: 0;
 
     @media (min-width: 768px) {
       margin-left: auto;
     }
   }
-  & .MuiTablePaginationUnstyled-actions {
-    padding: 2px;
-    border: 1px solid ${theme.palette.mode === 'dark' ? grey[800] : grey[200]};
-    border-radius: 50px;
-    text-align: center;
+
+  & .${classes.spacer} {
+    display: none;
   }
-  & .MuiTablePaginationUnstyled-actions > button {
-    margin: 0 8px;
-    border: transparent;
-    border-radius: 2px;
-    background-color: transparent;
-    &:hover {
-      background-color: ${theme.palette.mode === 'dark' ? grey[800] : grey[50]};
-    }
-    &:focus {
-      outline: 1px solid ${theme.palette.mode === 'dark' ? blue[400] : blue[200]};
-    }
+
+  & .${classes.actions} {
+    display: flex;
+    gap: 0.25rem;
   }
-  `,
-);
+`;
 
 export default function UnstyledTable() {
   const [page, setPage] = React.useState(0);
@@ -144,7 +101,7 @@ export default function UnstyledTable() {
   };
 
   return (
-    <Root sx={{ width: 500, maxWidth: '100%' }}>
+    <Root sx={{ maxWidth: '100%', width: 500 }}>
       <table aria-label="custom pagination table">
         <thead>
           <tr>
@@ -160,15 +117,14 @@ export default function UnstyledTable() {
           ).map((row) => (
             <tr key={row.name}>
               <td>{row.name}</td>
-              <td style={{ width: 120 }} align="right">
+              <td style={{ width: 160 }} align="right">
                 {row.calories}
               </td>
-              <td style={{ width: 120 }} align="right">
+              <td style={{ width: 160 }} align="right">
                 {row.fat}
               </td>
             </tr>
           ))}
-
           {emptyRows > 0 && (
             <tr style={{ height: 41 * emptyRows }}>
               <td colSpan={3} />
