+++ docs/data/base/components/table-pagination/table-pagination.md
@@ -1,0 +1,20 @@
+---
+product: base
+title: React Table pagination component
+components: TablePaginationUnstyled
+packageName: '@mui/base'
+---
+
+# Unstyled table pagination
+
+<p class="description">Table pagination lets users navigate through data in tables.</p>
+
+{{"component": "modules/components/ComponentLinkHeader.js", "design": false}}
+
+## Basic TablePaginationUnstyled
+
+{{"demo": "TableUnstyled.js"}}
+
+## Customized look and feel
+
+{{"demo": "TableCustomized.js"}}
