--- docs/pages/api-docs/table-pagination-unstyled.json
@@ -1,48 +1,0 @@
-{
-  "props": {
-    "count": { "type": { "name": "number" }, "required": true },
-    "onPageChange": { "type": { "name": "func" }, "required": true },
-    "page": { "type": { "name": "custom", "description": "integer" }, "required": true },
-    "rowsPerPage": { "type": { "name": "custom", "description": "integer" }, "required": true },
-    "component": { "type": { "name": "elementType" } },
-    "components": {
-      "type": {
-        "name": "shape",
-        "description": "{ Actions?: elementType, DisplayedRows?: elementType, MenuItem?: elementType, Root?: elementType, Select?: elementType, SelectLabel?: elementType, Spacer?: elementType, Toolbar?: elementType }"
-      },
-      "default": "{}"
-    },
-    "componentsProps": {
-      "type": {
-        "name": "shape",
-        "description": "{ actions?: object, displayedRows?: object, menuItem?: object, root?: object, select?: object, selectLabel?: object, spacer?: object, toolbar?: object }"
-      },
-      "default": "{}"
-    },
-    "getItemAriaLabel": {
-      "type": { "name": "func" },
-      "default": "function defaultGetAriaLabel(type: ItemAriaLabelType) {\n  return `Go to ${type} page`;\n}"
-    },
-    "labelDisplayedRows": {
-      "type": { "name": "func" },
-      "default": "function defaultLabelDisplayedRows({ from, to, count }: LabelDisplayedRowsArgs) {\n  return `${from}–${to} of ${count !== -1 ? count : `more than ${to}`}`;\n}"
-    },
-    "labelRowsPerPage": { "type": { "name": "node" }, "default": "'Rows per page:'" },
-    "onRowsPerPageChange": { "type": { "name": "func" } },
-    "rowsPerPageOptions": {
-      "type": {
-        "name": "arrayOf",
-        "description": "Array&lt;number<br>&#124;&nbsp;{ label: string, value: number }&gt;"
-      },
-      "default": "[10, 25, 50, 100]"
-    }
-  },
-  "name": "TablePaginationUnstyled",
-  "styles": { "classes": [], "globalClasses": {}, "name": null },
-  "spread": true,
-  "forwardsRefTo": "HTMLTableCellElement",
-  "filename": "/packages/mui-base/src/TablePaginationUnstyled/TablePaginationUnstyled.tsx",
-  "inheritance": null,
-  "demos": "<ul><li><a href=\"/components/tables/\">Tables</a></li></ul>",
-  "cssComponent": false
-}
